Timer fixed for CC430. Issues to look into: Gateway example reboot by watchdog. Undifined reference to 'write_byte_to_flash' in fs.c

+++ d7aoss/framework/timer.h
@@ -47,7 +47,7 @@ static bool timer_add_event_in_stack( timer_event new_event );
 static bool timer_update_stack( void );
 
 void timer_wait_done( void );
-void timer_wait_ms( uint16_t ms );
+void timer_wait_ms( uint32_t ms );
 
 void benchmarking_timer_init();
 uint32_t benchmarking_timer_getvalue();
